refactor product model and DTO to use string types for quantity, price, and profit percentage; add AllProductMapper for product transformation

+++ src/modules/ProductBarber/adapters/http/controller/productController.ts
@@ -1,11 +1,9 @@
 import { Request, Response } from "express";
 import { ProductService } from "../../../core/services/productService";
 import { IProduct } from "../../../core/domain/IProduct";
-import e = require("express");
-import { plainToClass } from "class-transformer";
-import ProductCreateDTO from "../../dtos/ProductCreateDTO";
-import { validate, ValidationError } from "class-validator";
 import { errorHandler } from "../../middleware/error/error";
+import { AllProductMapper } from "../../../core/mapper/AllProductMapper";
+import { deleteBlob, downloadBlob, uploadBlob } from "../../../../../adapters/azure/services/blobService";
 
 
 export class ProductController {
@@ -16,8 +14,8 @@ export class ProductController {
 
     async findAll(req: Request, res: Response) {
         try {
-            const products = await this.productService.findAll();
-            return res.json(products);
+            const products = await this.productService.findAll();            
+            return res.json(products.map(AllProductMapper));
         } catch (error) { 
             return res.status(500).json({ message: 'Internal server error' });
         }
@@ -32,16 +30,45 @@ export class ProductController {
             const product = await this.productService.findById(id);
             if (!product) {
                 return res.status(404).json({ message: 'Product not found' });
-            }
+            }          
             return res.json(product);
         } catch (error) {
             return res.status(500).json({ message: 'Internal server error' });
         }
     }
 
-    async create(req: Request, res: Response) {
+    async findImage(req: Request, res: Response) {
+        const { id } = req.params;
+        try {
+            if (!id) {
+                return res.status(400).json({ message: 'Id is required' });
+            }
+            const product = await this.productService.findById(id);
+            if (!product) {
+                return res.status(404).json({ message: 'Product not found' });
+            } 
+            const imageBuffer = await downloadBlob(product.urlImage);            
+            res.set('Content-Type', 'image/jpeg'); // Defina o tipo de conteúdo correto
+            res.send(imageBuffer);            
+        } catch (error) {
+            return res.status(500).json({ message: 'Internal server error' });
+        }
+    }
+
+    async create(req: Request, res: Response) {        
         const product: IProduct = req.body;
-        try {                       
+        try { 
+            if (!req.file) {
+                res.status(400).send("Nenhum arquivo enviado.");
+                return;
+            }
+            const file = req.file;
+            const blobUrl = await uploadBlob(file.originalname, file.buffer);    
+            if (!blobUrl) {
+                res.status(500).send("Erro ao fazer upload do arquivo.");
+                return;
+            } 
+            product.urlImage = blobUrl;                
             const newProduct = await this.productService.create(product);
             return res.json(newProduct);
         } catch (error) {               
@@ -62,6 +89,17 @@ export class ProductController {
     async delete(req: Request, res: Response) {
         const { id } = req.params;
         try {
+            if (!id) {
+                console.log('Id is required');
+                return res.status(500).json({ message: 'Internal server error' });
+            }
+            const product = await this.productService.findById(id);
+            
+            if (!product) {
+                console.log('Product not found');
+                return res.status(500).json({ message: 'Internal server error' });
+            }            
+            await deleteBlob(product.urlImage);           
             await this.productService.delete(id);
             return res.status(204).send();
         } catch (error) {
